requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
afk-parser @ git+https://github.com/astr0n0mer/afk_parser.git@stable
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
babel==2.16.0
    # via -r requirements.in
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via pymongo
fastapi==0.112.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   requests
motor==3.5.1
    # via -r requirements.in
parsedatetime==2.6
    # via afk-parser
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pymongo==4.8.0
    # via motor
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.9
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
slack-sdk==3.31.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.38.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via requests
uvicorn==0.30.6
    # via -r requirements.in
